Split builtin file into different files
